move base test suite class file to appropriate directory
