+++ release_tester/license_manager_tests/cluster.py
@@ -2,7 +2,7 @@
 
 # pylint: disable=import-error
 from license_manager_tests.base.cluster_base import LicenseManagerClusterBaseTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 
 
 class LicenseManagerClusterTestSuite(LicenseManagerClusterBaseTestSuite):
