+++ release_tester/package_installation_tests/community_package_installation_test_suite.py
@@ -11,7 +11,7 @@ from package_installation_tests.installation_steps import (
     check_if_client_package_can_be_installed,
     check_if_server_package_can_be_installed,
 )
-from selenium_ui_test.test_suites.base_test_suite import (
+from test_suites_core.base_test_suite import (
     testcase,
     disable_if_returns_true_at_runtime,
     disable_for_debian,
