+++ release_tester/selenium_ui_test/test_suites/activefailover/wait_for_upgrade_test_suite.py
@@ -2,7 +2,7 @@
 """active failover upgrade testsuite"""
 
 from selenium_ui_test.test_suites.activefailover.active_failover_base_suite import ActiveFailoverBaseTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 
 
 class ActiveFailoverWaitForUpgradeTestSuite(ActiveFailoverBaseTestSuite):
