+++ release_tester/selenium_ui_test/test_suites/analyzers_test_suite.py
@@ -4,7 +4,7 @@ import semver
 
 from selenium_ui_test.pages.analyzers_page import AnalyzerPage
 from selenium_ui_test.test_suites.base_selenium_test_suite import BaseSeleniumTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 
 
 class AnalyzersTestSuite(BaseSeleniumTestSuite):
