+++ release_tester/selenium_ui_test/test_suites/cluster/jam_2_test_suite.py
@@ -5,7 +5,7 @@ from selenium_ui_test.pages.nodes_page import NodesPage
 from selenium_ui_test.pages.cluster_page import ClusterPage
 from selenium_ui_test.pages.navbar import NavigationBarPage
 from selenium_ui_test.test_suites.base_selenium_test_suite import BaseSeleniumTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 
 
 class ClusterJamStepTwoSuite(BaseSeleniumTestSuite):
