+++ release_tester/selenium_ui_test/test_suites/graph_test_suite.py
@@ -2,7 +2,7 @@
 """ graph testsuite """
 
 from selenium_ui_test.test_suites.base_selenium_test_suite import BaseSeleniumTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 from selenium_ui_test.pages.graph_page import GraphPage, GraphExample, get_graph
