+++ release_tester/selenium_ui_test/test_suites/leader_follower/after_install_test_suite.py
@@ -4,7 +4,7 @@ import time
 
 from selenium_ui_test.pages.navbar import NavigationBarPage
 from selenium_ui_test.pages.replication_page import ReplicationPage
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 from selenium_ui_test.test_suites.base_classes.after_install_test_suite import AfterInstallTestSuite
