+++ release_tester/selenium_ui_test/test_suites/leader_follower/jam_1_test_suite.py
@@ -2,7 +2,7 @@
 """ jamming leader follower testsuite """
 
 from selenium_ui_test.test_suites.base_selenium_test_suite import BaseSeleniumTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 
 
 from selenium_ui_test.pages.navbar import NavigationBarPage
