+++ release_tester/selenium_ui_test/test_suites/query_test_suite.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 """ query testsuite """
 from selenium_ui_test.test_suites.base_selenium_test_suite import BaseSeleniumTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 from selenium_ui_test.pages.query_page import QueryPage
 from selenium_ui_test.pages.graph_page import GraphPage
 from selenium_ui_test.pages.graph_page import GraphExample, get_graph
