+++ release_tester/selenium_ui_test/test_suites/service_test_suit.py
@@ -2,7 +2,7 @@
 """ service page testsuite """
 from selenium_ui_test.pages.service_page import ServicePage
 from selenium_ui_test.test_suites.base_selenium_test_suite import BaseSeleniumTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 
 class ServiceTestSuite(BaseSeleniumTestSuite):
     """ analyzer page testsuite """
