+++ release_tester/selenium_ui_test/test_suites/support_test_suite.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 """ support page testsuite """
 from selenium_ui_test.test_suites.base_selenium_test_suite import BaseSeleniumTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 from selenium_ui_test.pages.support_page import SupportPage
