+++ release_tester/selenium_ui_test/test_suites/views_test_suite.py
@@ -1,9 +1,8 @@
 #!/usr/bin/env python3
 """ views testsuite """
-import semver
 
 from selenium_ui_test.test_suites.base_selenium_test_suite import BaseSeleniumTestSuite
-from selenium_ui_test.test_suites.base_test_suite import testcase
+from test_suites_core.base_test_suite import testcase
 from selenium_ui_test.pages.views_page import ViewsPage
 import time
 import semver
